Changed _accumulator to double from int

+++ Calculator/Calculator.cs
@@ -52,7 +52,7 @@ namespace Calculator
             return _accumulator;
         }
 
-        private int _accumulator;
+        private double _accumulator;
 
     }
 }
